fix(ui) Fix graph flicker on tab change
Because the activeTab state was held higher up in the render tree, when
tabs were changed the graph would re-render causing it to flicker.
Extracting a stateful component closer to the tree that changes fixes
that problem.
